Refactor CodeEditor init of editor
